requirements.txt
streamlit
google-generativeai
python-dotenv
panel
django
